renamed get/setDataObj to get/setDataObject
renamed submitCaption to submitValue

changed jala.Form.createComponents to use a stripped down label text if no name is provided for a component.

The constructor of a component now throws an error if no argument is provided. jala.Form.create writes that error to the log.

renamed jala.Form.REQUIRED to jala.Form.REQUIRE 



git-svn-id: http://dev.orf.at/source/jala/trunk@231 f3d1f554-2238-0410-8348-9ec7ac89951c