introduced getter and setter field for instances of jala.Form. these functions are used if no getter/setter is provided for a component. By default they use the static jala.Form.propertyGetter/Setter which return fields of the data object.
fixed date component: default value is used only if it really is a date, all other data types are omitted.

removed Component.File#setValue again. It's ok to force users to add a setter method (actually thought this method would fix a bug, but that turned out not to be a bug at all).


git-svn-id: http://dev.orf.at/source/jala/trunk@233 f3d1f554-2238-0410-8348-9ec7ac89951c